requirements.txt
scikit-learn~=1.1.3
numpy~=1.23.5
scipy~=1.9.3
nltk~=3.7
torchtext~=0.14.1
tqdm~=4.64.1
gensim~=4.2.0